+++ app/src/main/java/gmail/ahmedmeabbas/realestateapp/listings/addlisting/domain/AddApartmentUseCase.kt
@@ -11,15 +11,15 @@ class AddApartmentUseCase(
 
     fun addApartmentDetails(
         area: Double?,
-        yearBuilt: Int?,
+        yearBuilt: Long?,
         finished: Boolean?,
         furnished: Boolean?,
-        bedrooms: Int?,
-        bathrooms: Int?,
+        bedrooms: Long?,
+        bathrooms: Long?,
         kitchen: Boolean?,
         livingRoom: Boolean?,
         balcony: Boolean?,
-        floorNumber: Int?,
+        floorNumber: Long?,
         moreInfo: String?
     ) {
         apartment.apply {
