+++ app/src/main/java/gmail/ahmedmeabbas/realestateapp/listings/addlisting/presentation/house/HouseDetailsViewModel.kt
@@ -15,16 +15,16 @@ class HouseDetailsViewModel @Inject constructor(
     fun getPropertyType() = addListingRepository.getPropertyType()
 
     fun addHouseDetails(
-        bedrooms: Int?,
-        bathrooms: Int?,
-        kitchens: Int?,
-        halls: Int?,
-        floors: Int?,
+        bedrooms: Long?,
+        bathrooms: Long?,
+        kitchens: Long?,
+        halls: Long?,
+        floors: Long?,
         apartments: Boolean?,
-        numberApartments: Int?,
+        numberApartments: Long?,
         basement: Boolean?,
         garage: Boolean?,
-        numberCars: Int?,
+        numberCars: Long?,
         moreInfo: String?
     ) {
         addHouseUseCase.addHouseDetails(
